+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * Parsing array query parameters
 * slow `/v3/transactions` get requests (same as 3.5.3)
+* incorrect `/v3/asset_pairs` filtering
 
 ## 3.7.2