+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * slow `/v3/transactions` get requests (same as 3.5.3)
 * incorrect `/v3/asset_pairs` filtering
 * panic on `/v3/history`
+* empty relationships in create asset request
 
 ## 3.7.2