+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ## Added
 
 * Endpoint for getting account list `v3/accounts`
+* Filter by asset codes (`v3/assets`)
+* Filter by sales ids (`v3/sales`)
 
 ## Fixed